# PR: Migrated to Vite, Revamped UI & Improved Security