--- a/.env.example
+++ b/.env.example
@@ -1,1 +1,5 @@
-VITE_YOUTUBE_API_KEY=
+# To get your own API key, go to https://console.cloud.google.com/apis/api/youtube.googleapis.com/
+# Create a new project and enable the YouTube Data API v3
+# Then create an API key and add it to the .env file
+
+VITE_YOUTUBE_API_KEY=your_api_key