--- a/README.md
+++ b/README.md
@@ -1,62 +1,84 @@
-# Distraction-Free YouTube Player
-![YouTube Logo](https://upload.wikimedia.org/wikipedia/commons/b/b8/YouTube_Logo_2017.svg)
-Welcome to the **Distraction-Free YouTube Player**, a platform developed for focused and optimized learning with YouTube videos!
+# Distraction Free YouTube Player
+
+<div align="center">
+  <img src="https://upload.wikimedia.org/wikipedia/commons/b/b8/YouTube_Logo_2017.svg" alt="YouTube Logo" width="200px" />
+  <br/>
+  <img src="demo.png" alt="Demo Screenshot" width="100%" />
+</div>
+
+Welcome to the **Distraction Free YouTube Player**, a platform developed for focused and optimized learning with YouTube videos!
+
+## Platform Features
+
+- **Ad-Free Experience**: Watch tutorials without any advertisements
+- **Curated Content**: Handpicked programming tutorials from top educators
+- **Open Source**: Community-driven development, free for everyone
+- **Distraction-Free**: Clean interface without recommendations or comments
+- **Multiple Playlists**: Organized learning paths for different technologies
+- **Saved Progress**: Resume where you left off with automatic progress tracking
+- **Focused Learning**: The interface is tailored for maximum concentration
+- **Optimized Performance**: Fast loading and smooth playback
 
 
+## Tech Stack
 
-## Features:
-- **Distraction-Free YouTube Learning**: Watch YouTube videos in a clean, minimalist interface without distractions.
-- **2x Page Optimization Speed**: Improved speed for a seamless experience.
-- **Lite Version**: Designed to be light on resources, allowing you to focus on learning.
-- **Focused Learning**: The interface is tailored for maximum concentration.
+- React.js
+- Tailwind CSS
+- YouTube Data API
+- Framer Motion
+- React Player
 
----
 
-## Setup Guide:
+## Setup Guide
+
 To get started with the **Distraction-Free YouTube Player**, follow these steps:
 
-1. **Navigate to the frontend directory**:
+1. **Clone the repository**:
    ```bash
-   cd frontend
+   git clone https://github.com/hotheadhacker/youtube-player.git
+   cd youtube-player
    ```
 
-2. **Install the dependencies**:
+2. **Copy the example environment file**:
+   ```bash
+   cp .env.example .env
+   ```
+
+3. **Add your API key**:
+   Open the `.env` file and replace `your_api_key` with your actual YouTube API key.
+
+4. **Install the dependencies**:
    ```bash
    npm install
    ```
 
-3. **Run the development server**:
+5. **Run the development server**:
    ```bash
-   npm start
+   npm run dev
    ```
 
-   This will start the application locally. Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
+   This will start the application locally. Open [http://localhost:5173](http://localhost:5173) to view it in your browser.
 
----
+## Contributing
 
-## Development Note:
-This project is currently in the development phase. At this stage, the code may include hardcoded elements and can be challenging for beginners to understand. However, we have plans to refactor the frontend code to make it more readable, with clean and modular components for improved maintainability.
+We welcome contributions to make this platform even better! Here's how you can help:
 
----
+1. Fork the repository
+2. Create your feature branch (`git checkout -b feature/AmazingFeature`)
+3. Commit your changes (`git commit -m 'Add some AmazingFeature'`)
+4. Push to the branch (`git push origin feature/AmazingFeature`)
+5. Open a Pull Request
 
-## Future Goals:
-- **User Base Personalization (Stateful)**: Customization based on user preferences (will require cookies).
-- **Dedicated Account for Each User**: Sign in and have your own personalized space.
-- **Course Completion Tracking**: Track the progress of your courses and videos.
-- **Categories**: Organize your learning into various categories (Ideas are welcome!).
+Please ensure your PR description clearly describes the changes and their benefits.
 
----
+## Credits
 
-## Source Code:
-Check out the source code on GitHub:
-
-[hotheadhacker/youtube-player](https://github.com/hotheadhacker/youtube-player)
-
-
-
-## Credits:
 Platform designed by [Salman Qureshi](https://isalman.dev) and redesigned by [Zaid Adil](https://hothead01th.vercel.app).
 
+## Source Code
 
+Check out the source code on GitHub: [hotheadhacker/youtube-player](https://github.com/hotheadhacker/youtube-player)
 
-Thank you for trying the Distraction-Free YouTube Player! Stay tuned for updates.
+---
+
+Thank you for using the Distraction-Free YouTube Player! If you find it helpful, please consider giving it a star ⭐