--- a/frontend/postcss.config.js
+++ b/frontend/postcss.config.js
@@ -1,7 +1,0 @@
-// frontend/postcss.config.js
-module.exports = {
-    plugins: {
-      tailwindcss: {},
-      autoprefixer: {},
-    },
-  }