--- a/frontend/src/index.css
+++ b/frontend/src/index.css
@@ -1,3 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;