--- a/src/App.css
+++ b/src/App.css
@@ -1,4 +1,4 @@
-
+@import "tailwindcss";
 .pointer {cursor: pointer;}
 .skill-img {
   /* padding: 50px; */