--- a/src/components/LandingPage.jsx
+++ b/src/components/LandingPage.jsx
@@ -262,7 +262,7 @@ export default function LandingPage() {
             >
               <div className="absolute inset-0 bg-gradient-to-r from-indigo-500 to-purple-500 rounded-lg" />
               <div className="px-8 py-4 bg-black rounded-[6px] relative group transition duration-200 text-white hover:bg-transparent">
-                Start Learning
+               Get Started
               </div>
             </Link>
           </motion.div>