--- a/src/components/LandingPage.jsx
+++ b/src/components/LandingPage.jsx
@@ -1,8 +1,11 @@
 "use client";
-import { cn } from "../lib/utils";
 import { motion, useMotionTemplate, useMotionValue } from "framer-motion";
-import React from "react";
 import { Link } from "react-router-dom";
+
+// Utils function included directly
+function cn(...inputs) {
+  return inputs.filter(Boolean).join(' ');
+}
 
 export const BackgroundLines = ({
   children,