--- a/src/components/LandingPage.jsx
+++ b/src/components/LandingPage.jsx
@@ -1,4 +1,3 @@
-"use client";
 import { motion, useMotionTemplate, useMotionValue } from "framer-motion";
 import { Link } from "react-router-dom";