--- a/src/index.css
+++ b/src/index.css
@@ -1,1 +1,37 @@
 @import "tailwindcss";
+
+/* Custom Scrollbar Styles */
+::-webkit-scrollbar {
+    width: 10px;
+  }
+  
+  ::-webkit-scrollbar-track {
+    background: #1a1a1a;
+    border-radius: 5px;
+  }
+  
+  ::-webkit-scrollbar-thumb {
+    background: linear-gradient(to bottom, #6366f1, #a855f7);
+    border-radius: 5px;
+    border: 2px solid #1a1a1a;
+  }
+  
+  ::-webkit-scrollbar-thumb:hover {
+    background: linear-gradient(to bottom, #4f46e5, #9333ea);
+  }
+  
+  /* Playlist Specific Scrollbar */
+  .playlist-scroll::-webkit-scrollbar {
+    width: 8px;
+  }
+  
+  .playlist-scroll::-webkit-scrollbar-track {
+    background: #1a1a1a;
+    border-radius: 4px;
+  }
+  
+  .playlist-scroll::-webkit-scrollbar-thumb {
+    background: linear-gradient(to bottom, #6366f1, #a855f7);
+    border-radius: 4px;
+    border: 2px solid #1a1a1a;
+  }