… demo.png
… frontend/.gitignore
… frontend/README.md
… frontend/build/asset-manifest.json
… frontend/build/index.html
… frontend/build/static/css/main.270e2a8f.chunk.css
… frontend/build/static/css/main.270e2a8f.chunk.css.map
… frontend/build/static/js/2.6e479174.chunk.js
… frontend/build/static/js/2.6e479174.chunk.js.LICENSE.txt
… frontend/build/static/js/2.6e479174.chunk.js.map
… frontend/build/static/js/3.38c5d8a1.chunk.js
… frontend/build/static/js/3.38c5d8a1.chunk.js.map
… frontend/build/static/js/main.26e40ca9.chunk.js
… frontend/build/static/js/main.26e40ca9.chunk.js.map
… frontend/build/static/js/runtime-main.fb5fb25b.js
… frontend/build/static/js/runtime-main.fb5fb25b.js.map
… frontend/build/static/media/deno.820d88a4.svg
… frontend/build/static/media/docker.1497612c.png
… frontend/build/static/media/graphql.9ec42864.png
… frontend/build/static/media/nodejs.4ffd30ab.png
… frontend/build/static/media/python.0e3d57bf.png
… frontend/build/static/media/reactjs.7927890c.jpg
… frontend/build/static/media/typescript.c2eef6df.png
… frontend/package.json
… frontend/public/favicon.ico
… frontend/public/logo192.png
… frontend/public/logo512.png
… frontend/public/manifest.json
… frontend/src/App.js
… frontend/src/index.js