What’s Changed?
- Migrated from CRA to Vite → Cold start `(25s → 2s`), hot reload `(<1s`), and build times `(60s → 15s)`.
- Revamped UI → More modern, scalable, and responsive.
- Removed deprecated packages → Better stability & security.
- Improved security → Moved hardcoded API keys to `.env`.
- Optimized assets → Replaced image logos with `React Icons` for better performance.
- Better error handling → Added `react-toastify` for cleaner alerts.
- Added regex validation → Improved link handling in description.
- The app is now faster, safer, and more scalable. 🚀
- live link : https://free-youtube-player.vercel.app/