@HOTHEAD01TH 2025-02-24
> @HOTHEAD01TH merged, added new url, also do we need to set up new deployment pipelines?

@hotheadhacker No additional changes to the pipeline are needed at the moment. However, since you're using Netlify, you might need to add the YOUTUBE_API_KEY in the Netlify environment variable dashboard and update the root directory from frontend to empty (since the main project is now the root dir).